feat: adiciona função para fechar modal
