+++ src/components/commons/Modal/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import styled, { createGlobalStyle, css } from 'styled-components';
 import PropTypes from 'prop-types';
 import { motion } from 'framer-motion';
-import { Button } from '../Button';
+import { CloseButton } from './CloseButton';
 
 const ModalWrapper = styled.div`
   display: flex;
@@ -57,19 +57,6 @@ function handleClose(e, onClose) {
   }
 }
 
-const CloseButtonWrapper = styled(Button)`
-  position: absolute;
-  top: 20px;
-  right: 20px;
-  height: 50px;
-  width: 50px;
-  border-radius: 50%;
-`;
-
-function CloseButton() {
-  return <CloseButtonWrapper>X</CloseButtonWrapper>;
-}
-
 export default function Modal({ isOpen, onClose, children }) {
   return (
     <ModalWrapper
@@ -97,7 +84,7 @@ export default function Modal({ isOpen, onClose, children }) {
           flex: 1,
         }}
       >
-        {children(CloseButton, { 'data-modal-safe-area': 'true' })}
+        {children(() => CloseButton(onClose), { 'data-modal-safe-area': 'true' })}
       </motion.div>
     </ModalWrapper>
   );
